requirements.txt
numpy>=1.24
matplotlib>=3.7
pandas>=2.0
pyyaml>=6.0
gymnasium>=0.29
stable-baselines3>=2.2
scipy>=1.10
tqdm>=4.66
rich>=13.7
pillow>=10.0

SciencePlots>=2.1
